Fix Login.java bug and change directory for UserData
